move quantstudio design structure to structure directory
