+++ src/allotropy/parsers/appbio_quantstudio_designandanalysis/structure/generic/creator.py
@@ -3,7 +3,7 @@ from abc import ABC, abstractmethod
 from allotropy.parsers.appbio_quantstudio_designandanalysis.appbio_quantstudio_designandanalysis_contents import (
     DesignQuantstudioContents,
 )
-from allotropy.parsers.appbio_quantstudio_designandanalysis.appbio_quantstudio_designandanalysis_structure import (
+from allotropy.parsers.appbio_quantstudio_designandanalysis.structure.generic.structure import (
     Data,
 )
