+++ src/allotropy/parsers/appbio_quantstudio_designandanalysis/structure/generic/structure.py
