+++ src/allotropy/parsers/appbio_quantstudio_designandanalysis/structure/relative_standard_curve/structure.py
@@ -5,7 +5,7 @@ import pandas as pd
 from allotropy.parsers.appbio_quantstudio_designandanalysis.appbio_quantstudio_designandanalysis_contents import (
     DesignQuantstudioContents,
 )
-from allotropy.parsers.appbio_quantstudio_designandanalysis.appbio_quantstudio_designandanalysis_structure import (
+from allotropy.parsers.appbio_quantstudio_designandanalysis.structure.generic.structure import (
     WellList,
 )
