+++ tests/parsers/appbio_quantstudio_designandanalysis/appbio_quantstudio_designandanalysis_structure_test.py
@@ -8,7 +8,7 @@ from allotropy.exceptions import AllotropeConversionError
 from allotropy.parsers.appbio_quantstudio_designandanalysis.appbio_quantstudio_designandanalysis_contents import (
     DesignQuantstudioContents,
 )
-from allotropy.parsers.appbio_quantstudio_designandanalysis.appbio_quantstudio_designandanalysis_structure import (
+from allotropy.parsers.appbio_quantstudio_designandanalysis.structure.generic.structure import (
     Header,
     Result,
 )
